### 2024-05-13 #180 @rmj3197
QuadratiK Submission
Submitting Author: Raktim Mukhopadhyay (@rmj3197)
All current maintainers: @rmj3197, @giovsaraceno
Package Name: QuadratiK
One-Line Description of Package: QuadratiK includes test for multivariate normality, test for uniformity on the sphere, non-parametric two- and k-sample tests, random generation of points from the Poisson kernel-based density and clustering algorithm for spherical data.
Repository Link:  https://github.com/rmj3197/QuadratiK
Version submitted:   1.1.0
Archive: https://zenodo.org/records/14546750
Version accepted: 1.1.2

---

Documentation link : [https://quadratik.readthedocs.io/en/latest/](https://quadratik.readthedocs.io/en/latest/)

We introduce the `QuadratiK` package that incorporates innovative data analysis methodologies. The presented software, implemented in both `R` and `Python`, offers a comprehensive set of novel goodness-of-fit tests and clustering techniques using kernel-based quadratic distances. Our software implements one, two and k-sample tests for goodness of fit, providing an efficient and mathematically sound way to assess the fit of probability distributions. Expanded capabilities of our software include supporting tests for uniformity on the $d$-dimensional Sphere based on Poisson kernel densities, and algorithms for generating random samples from Poisson kernel densities. Particularly noteworthy is the incorporation of a unique clustering algorithm specifically tailored for spherical data that leverages a mixture of Poisson kernel-based densities on the sphere. Alongside this, our software includes additional graphical functions, aiding the users in validating, as well as visualizing and representing clustering results. This enhances interpretability and usability of the analysis. In summary, our `R` and `Python` packages serve as a powerful suite of tools, offering researchers and practitioners the means to delve deeper into their data, draw robust inference, and conduct potentially impactful analyses and inference across a wide array of disciplines.

If your package is associated with an existing community please check below:

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

    - The QuadratiK package offers robust tools for goodness-of-fit testing, a fundamental aspect in statistical analysis, where accurately assessing the fit of probability distributions is essential. This is especially critical in research domains where model accuracy has direct implications on conclusions and further research directions.
	    - Spherical data structures are common in fields such as biology, geosciences and astronomy, where data points are naturally mapped to a sphere. QuadratiK provides a tailored approach to effectively handle and interpret these data.
	    - This package is also of particular interest to professionals in health and biological sciences, where understanding and interpreting spherical data can be crucial in studies ranging from molecular biology to epidemiology and public health.

     - `SciPy` and `hyppo` also have collections of goodness-of-fit test functionalities. Our interest focuses on tests that are based on the family of kernel-based quadratic distances. The kernels we use are diffusion kernels, that is, probability distributions that depend on a tuning parameter and satisfy the convolution property. We also implement the Poisson kernel-based tests for uniformity on the d-dimensional sphere.

      - We are aware of only a limited number of `Python` libraries that offer spherical clustering capabilities, such as `spherecluster` (last updated in November 2018) and `soyclustering` (last updated in May 2020). `spherecluster` implements Spherical K-Means and clustering using von Mises Fisher distributions as proposed in ["Banerjee, Arindam, et al. "Clustering on the Unit Hypersphere using von Mises-Fisher Distributions." Journal of Machine Learning Research 6.9 (2005)."](http://www.jmlr.org/papers/volume6/banerjee05a/banerjee05a.pdf). `soyclustering` implements spherical k-means for document clustering which has been proposed in [Kim, Hyunjoong, Han Kyul Kim, and Sungzoon Cho. "Improving spherical k-means for document clustering: Fast initialization, sparse centroid projection, and efficient cluster labeling." Expert Systems with Applications 150 (2020): 113288.](https://www.sciencedirect.com/science/article/abs/pii/S0957417420301135)

      - In summary, there are fundamental differences between QuadratiK and existing packages that are as follows -

        - The GOF tests are U-statistics based on centered kernels. The concept and methodology of centering is unique to our methods and is not part of the methods appearing in existing packages.
        - An algorithm for connecting the tuning parameter with the statistical properties of the test, namely power and degrees of freedom (DOF) is provided. This feature differentiates our novel methods from methods in other packages.
        - A new clustering algorithm for data that reside on the sphere using the Poisson kernel-based densities is offered. This aspect is not a feature of the existing packages.
        - We also offer algorithms for generating random samples from Poisson kernel-based densities. This capability is also unique to our package.

      - We also implement a GUI to enable interaction with the software in a non-programmatic manner using the `streamlit` library. We have not found any python package that implements a GUI for the above described tasks.

    Please see our pre-submission enquiry for this submission at -
    https://github.com/pyOpenSci/software-submission/issues/168
