+++ QuadratiK/spherical_clustering/_pkbc.py
@@ -68,38 +68,59 @@ class PKBC:
 
     Attributes
     ----------
-        alpha\\_ : numpy.ndarray of shape (n_clusters,)
+        alpha\\_ : dict
             Estimated mixing proportions
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            and each value is a numpy.ndarray of shape (n_clusters,).
+            
 
-        labels\\_ : numpy.ndarray of shape (n_samples,)
+        labels\\_ : dict 
             Final cluster membership assigned by the algorithm to each observation
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            and each value is a numpy.ndarray of shape (n_samples,)
 
-        log_lik_vec : numpy.ndarray of shape (num_init, )
+        log_lik_vecs_ : dict
             Array of log-likelihood values for each initialization
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            and each value is a numpy.ndarray of shape (num_init, )
+                
 
-        loglik\\_ : float
+        loglik\\_ : dict
             Maximum value of the log-likelihood function
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            and each value is a float
+            
 
-        mu\\_ : numpy.ndarray of shape (n_clusters, n_features)
+        mu\\_ : dict
             Estimated centroids
-
-        num_iter_per_run : numpy.ndarray of shape (num_init, )
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            and each value is a numpy.ndarray of shape (n_clusters, n_features)
+            
+        num_iter_per_runs_ : dict
             Number of E-M iterations per run
-
-        post_probs\\_ : numpy.ndarray of shape (n_samples, n_features)
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            and each value is a numpy.ndarray of shape (num_init, )
+                
+        post_probs\\_ : dict
             Posterior probabilities of each observation for the indicated clusters
-
-        rho\\_ : numpy.ndarray of shape (n_clusters,)
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            and each value is a numpy.ndarray of shape (n_samples, n_features)
+            
+        rho\\_ : dict
             Estimated concentration parameters rho
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            and each value is a numpy.ndarray of shape (n_clusters,)
+            
+        euclidean\\_wcss\\_ : dict
+            Values of within-cluster sum of squares computed with Euclidean distance.
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            and each value is a float
 
-        euclidean\\_wcss\\_ : float
-            Values of within-cluster sum of squares computed with
-            Euclidean distance.
-
-        cosine\\_wcss\\_ : float
-            Values of within-cluster sum of squares computed with
-            cosine similarity.
-
+        cosine\\_wcss\\_ : dict
+            Values of within-cluster sum of squares computed with cosine similarity.
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            and each value is a float
+           
     References
     ----------
         Golzy M. & Markatou M. (2020) Poisson Kernel-Based
@@ -117,19 +138,6 @@ class PKBC:
     >>> le.fit(y)
     >>> y = le.transform(y)
     >>> cluster_fit = PKBC(num_clust=4, random_state=42).fit(X)
[…]
