+++ QuadratiK/__init__.py
@@ -1,6 +1,6 @@
 from importlib import metadata, import_module
 
-__version__ = "1.1.0"
+__version__ = "1.1.dev0"
 
 submodules = [
     "kernel_test",
