+++ doc/build/html/_sources/api_reference/generated/QuadratiK.kernel_test.select_h.rst
@@ -1,10 +1,0 @@
-﻿select\_h
-=========
-
-.. currentmodule:: QuadratiK.kernel_test
-
-.. autofunction:: select_h
-
-.. raw:: html
-
-     <div style='clear:both'></div>
