+++ doc/build/html/_sources/getting_started/installation.rst
@@ -1,36 +1,0 @@
-Installation
-=============
-
-.. title:: Getting Started : contents
-.. _installation:
-
-
-Which Python?
---------------
-
-You’ll need **Python 3.9 (except 3.9.7) or greater**.
-
-Install using ``pip``
------------------------
-``pip install QuadratiK``
-
-Dependencies
--------------
-QuadratiK requires the following (arranged alphabetically):
-
-- `matplotlib <https://matplotlib.org/>`_ (>=3.8.2)
-- `nest-asyncio <https://github.com/erdewit/nest_asyncio>`_ (>=1.5)
-- `numpy <https://numpy.org/>`_  (>= 1.26.2)
-- `pandas <https://pandas.pydata.org/docs/index.html>`_ (>= 2.1.3)
-- `plotly <https://plotly.com/python/>`_ (>=5.15.0)
-- `scikit-learn <https://scikit-learn.org/stable/>`_ (>= 1.3)
-- `scipy <https://docs.scipy.org/doc/scipy/reference/>`_ (>= 1.11)
-- `streamlit <https://streamlit.io/>`_ (>=1.30.0)
-- `tabulate <https://github.com/astanin/python-tabulate>`_ (>= 0.8)
-
-Testing
---------
-QuadratiK uses the Python ``pytest`` package.  
-To install ``pytest``, please go `here <https://docs.pytest.org/en/latest/getting-started.html#>`_.
-To run the tests using ``pytest``, please follow these `instructions <https://docs.pytest.org/en/latest/how-to/usage.html>`_.
-Navigate to the tests folder to run the tests. 
