+++ doc/build/html/_sources/user_guide/datasets.rst
@@ -1,3 +1,0 @@
-.. _datasets:
-
-.. include:: ../../../QuadratiK/datasets/data/wireless_localization_dataset.rst

[…]
