### 2024-05-17 f8595aa
updated docs and fix normality test
