### 2024-05-25 #180 @Batalex
**The package installation does not install the dependencies**

---
---

Nice submission, I'll get started on finding the perfect editor for `Quadratik`!
