+++ QuadratiK/spherical_clustering/_pkbc.py
@@ -71,57 +71,57 @@ class PKBC:
     ----------
         alpha\\_ : dict
             Estimated mixing proportions
-            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector,
             and each value is a numpy.ndarray of shape (n_clusters,).
-            
 
-        labels\\_ : dict 
+
+        labels\\_ : dict
             Final cluster membership assigned by the algorithm to each observation
-            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector,
             and each value is a numpy.ndarray of shape (n_samples,)
 
         log_lik_vecs_ : dict
             Array of log-likelihood values for each initialization
-            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector,
             and each value is a numpy.ndarray of shape (num_init, )
-                
+
 
         loglik\\_ : dict
             Maximum value of the log-likelihood function
-            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector,
             and each value is a float
-            
+
 
         mu\\_ : dict
             Estimated centroids
-            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector,
             and each value is a numpy.ndarray of shape (n_clusters, n_features)
-            
+
         num_iter_per_runs_ : dict
             Number of E-M iterations per run
-            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector,
             and each value is a numpy.ndarray of shape (num_init, )
-                
+
         post_probs\\_ : dict
             Posterior probabilities of each observation for the indicated clusters
-            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector,
             and each value is a numpy.ndarray of shape (n_samples, n_features)
-            
+
         rho\\_ : dict
             Estimated concentration parameters rho
-            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector,
             and each value is a numpy.ndarray of shape (n_clusters,)
-            
+
         euclidean\\_wcss\\_ : dict
             Values of within-cluster sum of squares computed with Euclidean distance.
-            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector,
             and each value is a float
 
         cosine\\_wcss\\_ : dict
             Values of within-cluster sum of squares computed with cosine similarity.
-            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector, 
+            A dictionary containing key-value pairs, where each key is an element from the `num_clust` vector,
             and each value is a float
-           
+
     References
     ----------
         Golzy M. & Markatou M. (2020) Poisson Kernel-Based
@@ -189,9 +189,9 @@ class PKBC:
                 raise Exception("Input parameter num_clust must be greater than 1")
             else:
[…]
