### 2024-05-27 60b61e1
Updated summary for GOF and included summary for PKBC, included __repr__, updated UI
