+++ QuadratiK/ui/pages/2_Two_Sample_Test.py
@@ -96,24 +96,26 @@ if (y_data) and (x_data) is not None:
 
         try:
             two_sample_test = run_twosample_test(h_val, num_iter, b, X, Y)
-            res = pd.DataFrame()
-            res["Value"] = [
-                two_sample_test.test_type_,
-                two_sample_test.un_test_statistic_,
-                two_sample_test.un_cv_,
-                two_sample_test.un_h0_rejected_,
-                two_sample_test.var_un_,
+            index_labels = [
+                "Test Statistic",
+                "Critical Value",
+                "H0 is rejected (1 = True, 0 = False)",
             ]
-            res = res.set_axis(
-                [
-                    "Test Type",
-                    "Un Test Statistic",
-                    "Un Critical Value",
-                    "Un Reject H0",
-                    "Var Un",
-                ]
-            )
+            test_summary = {
+                "Dn": [
+                    two_sample_test.dn_test_statistic_,
+                    two_sample_test.dn_cv_,
+                    two_sample_test.dn_h0_rejected_,
+                ],
+                "Trace": [
+                    two_sample_test.trace_test_statistic_,
+                    two_sample_test.trace_cv_,
+                    two_sample_test.trace_h0_rejected_,
+                ],
+            }
+            res = pd.DataFrame(test_summary, index=index_labels)
 
+            st.text(two_sample_test.test_type_)
             st.table(res)
             csv_res = res.to_csv().encode()
             st.download_button(
