### 2024-05-27 95292c3
Update test_kernel_test.py
