### 2024-05-27 da2d850
Update test_kernel_test.py
