### 2024-05-27 bca2012
Update test_kernel_test.py
