+++ tests/test_kernel_test.py
@@ -23,17 +23,17 @@ class TestKernelTest(unittest.TestCase):
         dataframe_x = pd.DataFrame(x)
         normality_test_numpy = KernelTest(h=0.5, random_state=42).test(x)
         normality_test_dataframe = KernelTest(h=0.5, random_state=42).test(dataframe_x)
-        self.assertFalse(normality_test_numpy.dn_h0_rejected_)
+        self.assertFalse(normality_test_numpy.un_h0_rejected_)
         self.assertTrue(
             isinstance(normality_test_numpy.un_test_statistic_, (int, float))
         )
-        self.assertFalse(normality_test_dataframe.dn_h0_rejected_)
+        self.assertFalse(normality_test_dataframe.un_h0_rejected_)
         self.assertTrue(
             isinstance(normality_test_dataframe.un_test_statistic_, (int, float))
         )
         self.assertEqual(
-            normality_test_numpy.dn_h0_rejected_,
-            normality_test_dataframe.dn_h0_rejected_,
+            normality_test_numpy.un_h0_rejected_,
+            normality_test_dataframe.un_h0_rejected_,
         )
         self.assertEqual(
             normality_test_numpy.un_test_statistic_,
@@ -47,7 +47,7 @@ class TestKernelTest(unittest.TestCase):
         normality_test_1d = KernelTest(
             method="subsampling", b=0.5, random_state=42, alternative="scale"
         ).test(x_1d)
-        self.assertFalse(normality_test_1d.dn_h0_rejected_)
+        self.assertFalse(normality_test_1d.un_h0_rejected_)
         self.assertTrue(isinstance(normality_test_1d.un_test_statistic_, (int, float)))
 
     def test_twosample(self):
@@ -60,7 +60,7 @@ class TestKernelTest(unittest.TestCase):
         ).test(x, y)
         self.assertTrue(np.all(two_sample_test_subsampling.dn_h0_rejected_))
         self.assertTrue(
-            isinstance(two_sample_test_subsampling.dn_test_statistic_, np.ndarray)
+            isinstance(two_sample_test_subsampling.dn_test_statistic_, (int,float))
         )
 
         with self.assertRaises(ValueError):
@@ -84,7 +84,7 @@ class TestKernelTest(unittest.TestCase):
         ).test(x, y)
         self.assertTrue(np.all(two_sample_test_h_selection.dn_h0_rejected_))
         self.assertTrue(
-            isinstance(two_sample_test_h_selection.dn_test_statistic_, np.ndarray)
+            isinstance(two_sample_test_h_selection.dn_test_statistic_, (int, float))
         )
 
     def test_ksample(self):
@@ -98,21 +98,21 @@ class TestKernelTest(unittest.TestCase):
             h=1.5, method="subsampling", b=0.5, random_state=42
         ).test(x, y)
         self.assertFalse(np.all(k_sample_test.dn_h0_rejected_))
-        self.assertTrue(isinstance(k_sample_test.dn_test_statistic_, np.ndarray))
+        self.assertTrue(isinstance(k_sample_test.dn_test_statistic_, (int, float)))
 
         k_sample_test_without_h = KernelTest(
             method="subsampling", b=0.5, random_state=42, alternative="location"
         ).test(x, y)
         self.assertFalse(np.all(k_sample_test_without_h.dn_h0_rejected_))
         self.assertTrue(
-            isinstance(k_sample_test_without_h.dn_test_statistic_, np.ndarray)
+            isinstance(k_sample_test_without_h.dn_test_statistic_, (int, float))
         )
 
         k_sample_test_1d = KernelTest(
             h=1.5, method="subsampling", b=0.5, random_state=42
[…]
