### 2024-05-28 8107f14
Updated basic_usage and included binder requirements.txt file
