+++ .binder/requirements.txt
@@ -1,0 +1,1 @@
+git+https://github.com/rmj3197/QuadratiK.git
