+++ QuadratiK/datasets/__init__.py
@@ -1,3 +1,7 @@
-from ._dataset import load_wireless_data
+from ._dataset import (
+    load_wireless_data,
+    load_wisconsin_breast_cancer_data,
+    load_wine_data,
+)
 
-__all__ = ["load_wireless_data"]
+__all__ = ["load_wireless_data", "load_wisconsin_breast_cancer_data", "load_wine_data"]
