+++ QuadratiK/datasets/_dataset.py
@@ -107,3 +107,260 @@ def load_wireless_data(desc=False, return_X_y=False, as_dataframe=True, scaled=F
             return pd.DataFrame(data, columns=feature_names)
     else:
         return data
+
+
+def load_wisconsin_breast_cancer_data(
+    desc=False, return_X_y=False, as_dataframe=True, scaled=False
+):
+    """
+    The wireless data frame has 569 rows and 31 columns. The first 30 variables
+    report the features are computed from a digitized image of a fine needle aspirate (FNA) of a breast mass.
+    They describe characteristics of the cell nuclei present in the image.
+    The last column indicates the class labels (Benign = 0 or Malignant = 1).
+
+    The function load_wisconsin_breast_cancer_data loads the Breast Cancer Wisconsin (Diagnostic).
+
+    Read more in the :ref:`User Guide <datasets>`.
+
+    Parameters
+    ----------
+
+        desc : boolean, optional
+            If set to `True`, the function will return the description along with the data.
+            If set to `False`, the description will not be included. Defaults to False.
+
+        return_X_y : boolean, optional
+            Determines whether the function should return the data as separate arrays (`X` and `y`).
+            Defaults to False.
+
+        as_dataframe : boolean, optional
+            Determines whether the function should return the data as a pandas DataFrame (Trues)
+            or as a numpy array (False). Defaults to True.
+
+        scaled : boolean, optional
+            Determines whether or not the data should be scaled. If set to True, the data will be
+            divided by its Euclidean norm along each row. Defaults to False.
+
+    Returns
+    -------
+
+        (data, target) : tuple, if return_X_y is True
+            A tuple of two ndarray. The first containing a 2D array of shape
+            (n_samples, n_features) with each row representing one sample and
+            each column representing the features. The second ndarray of shape
+            (n_samples,) containing the target samples.
+
+        data : pandas.DataFrame, if as_dataframe is True
+            Dataframe of the data with shape (n_samples, n_features + class)
+
+        (desc, data, target) : tuple, if desc is True and return_X_y is True
+            A tuple of description and two numpy.ndarray. The first containing a 2D
+            array of shape (n_samples, n_features) with each row representing
+            one sample and each column representing the features. The second
+            ndarray of shape (n_samples,) containing the target samples.
+
+        (desc, data) : tuple, if desc is True and as_dataframe is True
+            A tuple of description and pandas.DataFrame.
+            Dataframe of the data with shape (n_samples, n_features + class)
+
+    References
+    ----------
+        Street, W. N., Wolberg, W. H., & Mangasarian, O. L. (1993, July).
+        Nuclear feature extraction for breast tumor diagnosis.
+        In Biomedical image processing and biomedical visualization (Vol. 1905, pp. 861-870). SPIE.
+
+    Source
+    -------
+        Wolberg,William, Mangasarian,Olvi, Street,Nick, and Street,W.. (1995).
+        Breast Cancer Wisconsin (Diagnostic). UCI Machine Learning Repository. https://doi.org/10.24432/C5DW2B.
+
+    Examples
+    --------
+    >>> from QuadratiK.datasets import load_wisconsin_breast_cancer_data
+    >>> X, y = load_wisconsin_breast_cancer_data(return_X_y=True)
+    """
+
+    data = np.loadtxt(
+        str(
+            resources.files("QuadratiK.datasets").joinpath(
+                "data/wisconsin_breast_cancer.txt"
+            )
+        )
+    )
+
+    if scaled:
+        data[:, :-1] = data[:, :-1] / np.linalg.norm(
+            data[:, :-1], axis=1, keepdims=True
+        )
+
[…]
