### 2024-05-29 aaa4d2b
Update _h_selection.py
