### 2024-05-29 86a63f2
Updated documentation, usage examples, added additional documentation for datasets, black formatting
