+++ QuadratiK/datasets/data/wine_dataset.rst
@@ -1,0 +1,45 @@
+.. _wine_dataset:
+
+Wine Dataset
+=============
+
+The `wine` data frame has 178 rows and 14 columns. The first 13 variables
+report 13 constituents found in each of the three types of wines.
+The last column indicates the class labels (1,2 or 3).
+
+Format
+------
+
+A data frame containing the following columns:
+
+- `Alcohol`
+- `Malic acid`
+- `Ash`
+- `Alcalinity of ash`
+- `Magnesium`
+- `Total phenols`
+- `Flavanoids`
+- `Nonflavanoid phenols`
+- `Proanthocyanins`
+- `Color intensity`
+- `Hue`
+- `OD280/OD315 of diluted wines`
+- `Proline`
+- `Class`
+
+Details
+-------
+
+These data are the results of a chemical analysis of wines grown in the same region in Italy but derived from three different cultivars. 
+The analysis determined the quantities of 13 constituents found in each of the three types of wines. 
+
+Source
+------
+
+Aeberhard,Stefan and Forina,M.. (1991). Wine. UCI Machine Learning Repository. https://doi.org/10.24432/C5PC7J.
+
+References
+----------
+
+Aeberhard, S., Coomans, D., & De Vel, O. (1994). Comparative analysis of statistical pattern recognition methods in high dimensional settings. Pattern Recognition, 27(8), 1065-1077.
+
