+++ doc/source/api_reference/generated/QuadratiK.kernel_test.KernelTest.rst
@@ -71,8 +71,6 @@
     
     
     
-    
-    
         KernelTest.stats
     
     
