+++ doc/source/user_guide/datasets.rst
@@ -1,3 +1,5 @@
 .. _datasets:
 
 .. include:: ../../../QuadratiK/datasets/data/wireless_localization_dataset.rst
+.. include:: ../../../QuadratiK/datasets/data/wisconsin_breast_cancer_dataset.rst
+.. include:: ../../../QuadratiK/datasets/data/wine_dataset.rst
