+++ doc/source/api_reference/generated/QuadratiK.ui.UI.rst
@@ -69,8 +69,6 @@
     
     
     
-    
-    
         UI.main
     
     
