+++ doc/source/user_guide/index.rst
@@ -15,4 +15,3 @@ Usage Examples
 
    basic_usage
    gen_plot_rpkb
-

[…]
