### 2024-05-29 bdf1ba6
Included Binder badge, updated usage example
