+++ doc/source/api_reference/index.rst
@@ -66,6 +66,8 @@ Datasets
    :toctree: generated/
 
    load_wireless_data
+   load_wisconsin_breast_cancer_data
+   load_wine_data
 
 .. automodule:: QuadratiK.tools
 
