+++ doc/source/api_reference/generated/QuadratiK.datasets.load_wine_data.rst
@@ -1,0 +1,10 @@
+﻿load\_wine\_data
+================
+
+.. currentmodule:: QuadratiK.datasets
+
+.. autofunction:: load_wine_data
+
+.. raw:: html
+
+     <div style='clear:both'></div>
