+++ doc/source/api_reference/generated/QuadratiK.datasets.load_wisconsin_breast_cancer_data.rst
@@ -1,0 +1,10 @@
+﻿load\_wisconsin\_breast\_cancer\_data
+=====================================
+
+.. currentmodule:: QuadratiK.datasets
+
+.. autofunction:: load_wisconsin_breast_cancer_data
+
+.. raw:: html
+
+     <div style='clear:both'></div>
