+++ doc/source/user_guide/datasets.rst
@@ -1,5 +1,16 @@
 .. _datasets:
 
+Wireless Localization Dataset
+-----------------------------
+
 .. include:: ../../../QuadratiK/datasets/data/wireless_localization_dataset.rst
+
+Wisconsin Breast Cancer Dataset
+-------------------------------
+
 .. include:: ../../../QuadratiK/datasets/data/wisconsin_breast_cancer_dataset.rst
+
+Wine Dataset
+------------
+
 .. include:: ../../../QuadratiK/datasets/data/wine_dataset.rst
