+++ doc/source/user_guide/index.rst
@@ -8,10 +8,16 @@ Dataset
 
    datasets.rst
 
-Usage Examples
----------------
+Common Usage Examples
+-----------------------
 .. toctree::
    :maxdepth: 4
 
    basic_usage
+
+Random Sampling from PKBD
+---------------------------
+.. toctree::
+   :maxdepth: 4
+
    gen_plot_rpkb
