### 2024-05-30 35c0442
updated PKBC docstring, and fixed Class datatype as int for datasets
