### 2024-05-30 06f5d5a
updated UI, included plot method for PKBC, include plot example, updated graphics.py
