### 2024-06-03 d997da4
included tests for additional datasets
