+++ tests/test_datasets.py
@@ -56,24 +56,42 @@ class TestLoadWirelessData(unittest.TestCase):
         )
 
     def test_desc(self):
-        descr, _ = load_wireless_data(desc=True)
-        self.assertTrue(isinstance(descr, str))
+        descr1, _ = load_wireless_data(desc=True)
+        self.assertTrue(isinstance(descr1, str))
+        descr2, _ = load_wisconsin_breast_cancer_data(desc=True)
+        self.assertTrue(isinstance(descr2, str))
+        descr3, _ = load_wine_data(desc=True)
+        self.assertTrue(isinstance(descr3, str))
 
     def test_scaled_as_numpy(self):
-        X, y = load_wireless_data(scaled=True, return_X_y=True)
-        self.assertEqual(X.shape[0], 2000)
-        self.assertEqual(y.shape[0], 2000)
-        self.assertTrue(isinstance(X, np.ndarray))
-        self.assertTrue(isinstance(y, np.ndarray))
+        X1, y1 = load_wireless_data(scaled=True, return_X_y=True)
+        X2, y2 = load_wisconsin_breast_cancer_data(scaled=True, return_X_y=True)
+        X3, y3 = load_wine_data(scaled=True, return_X_y=True)
+        self.assertTrue(isinstance(X1, np.ndarray))
+        self.assertTrue(isinstance(y1, np.ndarray))
+        self.assertTrue(isinstance(X2, np.ndarray))
+        self.assertTrue(isinstance(y2, np.ndarray))
+        self.assertTrue(isinstance(X3, np.ndarray))
+        self.assertTrue(isinstance(y3, np.ndarray))
 
     def test_desc_with_Xy(self):
-        descr, X, y = load_wireless_data(desc=True, return_X_y=True)
-        self.assertEqual(X.shape[0], 2000)
-        self.assertEqual(y.shape[0], 2000)
-        self.assertTrue(isinstance(descr, str))
+        descr1, X1, y1 = load_wireless_data(desc=True, return_X_y=True)
+        descr2, X2, y2 = load_wisconsin_breast_cancer_data(desc=True, return_X_y=True)
+        descr3, X3, y3 = load_wine_data(desc=True, return_X_y=True)
+        self.assertTrue(isinstance(descr1, str))
+        self.assertTrue(isinstance(descr2, str))
+        self.assertTrue(isinstance(descr3, str))
+        self.assertTrue(isinstance(X1, np.ndarray))
+        self.assertTrue(isinstance(y1, np.ndarray))
+        self.assertTrue(isinstance(X2, np.ndarray))
+        self.assertTrue(isinstance(y2, np.ndarray))
+        self.assertTrue(isinstance(X3, np.ndarray))
+        self.assertTrue(isinstance(y3, np.ndarray))
 
     def test_as_numpy(self):
-        df = load_wireless_data(as_dataframe=False)
-        self.assertEqual(df.shape[0], 2000)
-        self.assertEqual(df.shape[1], 8)
-        self.assertTrue(isinstance(df, np.ndarray))
+        df1 = load_wireless_data(as_dataframe=False)
+        df2 = load_wisconsin_breast_cancer_data(as_dataframe=False)
+        df3 = load_wine_data(as_dataframe=False)
+        self.assertTrue(isinstance(df1, np.ndarray))
+        self.assertTrue(isinstance(df2, np.ndarray))
+        self.assertTrue(isinstance(df3, np.ndarray))
