### 2024-06-04 c03f58f
updated jupyter notebooks
