+++ doc/source/api_reference/generated/QuadratiK.spherical_clustering.PKBC.rst
@@ -77,6 +77,8 @@
     
     
     
+    
+    
         PKBC.fit
     
     
