+++ doc/source/api_reference/generated/QuadratiK.kernel_test.KernelTest.rst
@@ -71,6 +71,8 @@
     
     
     
+    
+    
         KernelTest.stats
     
     
