+++ doc/source/api_reference/generated/QuadratiK.poisson_kernel_test.PoissonKernelTest.rst
@@ -71,6 +71,8 @@
     
     
     
+    
+    
         PoissonKernelTest.stats
     
     
