+++ doc/source/api_reference/generated/QuadratiK.ui.UI.rst
@@ -69,6 +69,8 @@
     
     
     
+    
+    
         UI.main
     
     
