+++ doc/source/api_reference/generated/QuadratiK.spherical_clustering.PKBD.rst
@@ -69,6 +69,8 @@
     
     
     
+    
+    
         PKBD.dpkb
     
     
