+++ doc/source/changelog/v1.1.1.rst
@@ -1,12 +1,8 @@
 QuadratiK Version 1.1.1
 =========================
 
-This version is currently in development.
-
-This version can be found on the master branch.
-
 This version includes - 
-    1. **[NEW]** Included repr method for KernelTest and PoissonKernelTest classes.
+    1. **[NEW]** Included `repr` method for KernelTest and PoissonKernelTest classes.
     2. **[NEW]** Separate attributes for Dn and Trace statistics, CV and H0.
     3. **[NEW]** Updated summary print format for KernelTest and PoissonKernelTest.
     4. **[NEW]** Updated Examples.
