+++ pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "QuadratiK"
-version = "1.1.dev0"
+version = "1.1.1"
 description = "QuadratiK includes test for multivariate normality, test for uniformity on the Sphere, non-parametric two- and k-sample tests, random generation of points from the Poisson kernel-based density and clustering algorithm for spherical data."
 license = "GPL-3.0-only"
 authors = [
