### 2024-07-01 #180 @isabelizimm
Welcome welcome to our fearless reviewers: @acolum and @ab93 👋  Thank you SO MUCH for volunteering to review for pyOpenSci! You are two people with awesome math-y, stats-y, ML-y, Python-y backgrounds, which is perfect for this package, and I am looking forward to learning from you through this review process 🌻

- [ ] @acolum survey completed.
- [x] @ab93 survey completed.

here as a comment, once your review is complete. You can look at other issues in this repository for examples of what this might look like.

Please get in touch with any questions or concerns! ~Your review is due 3 weeks from now, which is July 19.~ New review date: Aug 2. Please let me know if this date does not work for you.
---

tldr;
---
Reviewers: @acolum and @ab93
Due date [NOTE: deadline extended]: August 2
