### 2024-07-03 476b1a0
Bumped v1.1.1 -> v1.2.dev0
Bumped version to v1.2.dev0. Includes usage guide for Dashboard Application, minor updates to Normality Test in Dashboard Application and Tuning Parameter Selection.
