+++ doc/source/changelog/index.rst
@@ -6,6 +6,7 @@ Changelogs and release notes for all QuadratiK releases are linked in this page.
 .. toctree::
    :maxdepth: 1
 
+   v1.2.dev0.rst
    v1.1.1.rst
    v1.1.0.rst
    v1.0.0.rst
