### 2024-07-14 #180 @acolum
---

Overall, this submission was well done and followed most Python package development and documentation best practices. I found no major issues with the package's documentation, usability, and functionality, but I've outlined a few minor issues below.

Potential issues that could be fixed:
- The README file in the root directory of the package's repository has a .rst file extension instead of a .md file extension. Since I'm a first-time reviewer, I'm not sure how important this difference is to pyOpenSci.
- This wouldn't be required for final approval, but I think it would be helpful for potential users of your software to list similar packages in the R and Python scientific ecosystems in the README file. For example, what other packages implement similar methods or have similar overall functionality?

Minor issues that need fixing:
- The README file is missing a repo status badge. With the addition of this badge (and potentially others, like the pyOpenSci peer-review badge, in the future), I'd recommend organizing the badges in your README like the badges are organized [in this rOpenSci package's README](https://github.com/ropensci/drake).
- Although the README file links to the primary vignette with examples of how to use the package, the other vignettes in the documentation should be linked to here as well.
-  Currently, there are a few citations in the bibliography in the README file, but it's not clear which should be cited when citing the software package. More clarity and/or a BibTeX citation is needed. [This guidance from GitHub](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files) may help.
