### 2024-08-02 #180 @ab93
---

Great submission overall. Documentation is good, and I like the user guide as well.
There are a few tweaks, suggestions and adjustments that I can add here.

Packaging and CI

- The package uses the Poetry package manager, which is great. As a result, it would be nice to include a small guide to install the package for development using poetry.
e.g. `poetry install`
- In the package dependencies the Python version requirement is given as a broad `python = "^3.9, !=3.9.7"`.
This can cause some downstream applications using this package to break if any of its dependencies do not support a new Python version.
So my recommendation will be to only allow Python versions that the package supports, i.e. something like
`python = ">=3.9, <3.13"`
- It would be great to have some linting tools in the CI pipeline. Ruff is a great option, and will eliminate any anti-patterns in the coding practices
- I see the `Black` format badge, but I don't see the CI having the black format check. Adding that would be great

Code Practices, which again can be identified using a linter like Ruff

- Python typing is missing in the code, and would be a very nice to have, e.g. _pkbc.py
- Raising the base `Exception` class is not the best practice, a more fine-grained exception raising would be great
- A good practice is also to make sure all instance variables are first defined in the `__init__()` function, e.g. in the `PKBC` class,
`self.dat` only gets initialized in `fit()`
- Adding `__slots__` would be nice, as it reduces memory footprint
