### 2024-08-06 #180 @isabelizimm
Thank you so much to our reviewers @acolum and @ab93 for your thoughts on `QuadratiK`!🌷 The next step here is for the author to implement the changes suggested by reviewers. This piece can involve a bit of back and forth, @rmj3197, please let us know in this thread if you have questions about the review. Otherwise, post here when the reviews have been addressed and the reviewers will look over the updates and give their final approval!

> .rst file extension instead of a .md file extension.

This is okay! As long as there is a README file there, we are good to go 😄
