### 2024-09-01 f3ec67e
added ruff linting and additional workflows
