+++ QuadratiK/ui/pages/6_Data_generation_from_PKBD_Models.py
@@ -3,8 +3,6 @@ Contains the data generation from PKBD functionality of the UI
 """
 
 import importlib
-import copy
-from io import BytesIO
 import re
 import streamlit as st
 import pandas as pd
@@ -12,8 +10,6 @@ import numpy as np
 import matplotlib.pyplot as plt
 import plotly.graph_objects as go
 from plotly.subplots import make_subplots
-from sklearn.preprocessing import LabelEncoder
-from sklearn.utils.parallel import Parallel, delayed
 
 pkbd = importlib.import_module("QuadratiK.spherical_clustering").PKBD
 
