+++ QuadratiK/ui/pages/5_Uniformity_Test.py
@@ -45,7 +45,7 @@ header_exist = st.checkbox(
     "**Select, if the header is present in the data file.**", value=True
 )
 
-if header_exist == False:
+if header_exist is False:
     header = None
 else:
     header = "infer"
@@ -58,10 +58,8 @@ if data is not None:
     st.success(data.name + " Uploaded Successfully!")
     try:
         X = pd.read_csv(data, sep=delim, header=header)
-    except:
-        st.error(
-            "Unable to read the data file. Please make sure that the delimiter is correct."
-        )
+    except Exception as e:
+        st.error(f"An error occurred: {e}")
 
     num_iter = int(
         st.number_input(
@@ -116,8 +114,8 @@ if data is not None:
                 key="download-csv",
             )
             st.success("Done!")
-        except:
-            st.error("Please check user inputs and data file")
+        except Exception as e:
+            st.error(f"An error occurred: {e}")
 
 
 st.header("QQ Plots", divider="grey")
@@ -132,8 +130,8 @@ with st.expander("Click to view code"):
 if data is not None:
     try:
         st.pyplot(qq_plot(X, dist="uniform"), use_container_width=True)
-    except:
-        st.error("Please ensure that the data file is loaded")
+    except Exception as e:
+        st.error(f"An error occurred: {e}")
 
 
 st.markdown(
