### 2024-09-01 be31502
fixed linting in tests
