+++ tests/test_rpkb.py
@@ -3,8 +3,10 @@ Tests the PKBD class
 """
 
 import unittest
+
 import numpy as np
 import pandas as pd
+
 from QuadratiK.spherical_clustering import PKBD
 
 
