+++ tests/test_select_h.py
@@ -1,6 +1,8 @@
 import unittest
+
 import numpy as np
 import pandas as pd
+
 from QuadratiK.kernel_test import select_h
 
 
