+++ pyproject.toml
@@ -16,7 +16,7 @@ packages = [
 include = [{ path = "QuadratiK/ui/.streamlit", format = ["sdist", "wheel"] }]
 
 [tool.poetry.dependencies]
-python = "^3.9, !=3.9.7"
+python = ">=3.9, !=3.9.7, <3.13"
 scikit-learn = "^1.3"
 scipy = "^1.11"
 streamlit = "^1.30.0"
