### 2024-09-01 5727fcb
Ruff Linting
