+++ QuadratiK/poisson_kernel_test/_poisson_kernel_test.py
@@ -3,14 +3,14 @@ Poisson kernel-based quadratic distance test of Uniformity on the Sphere
 """
 
 import importlib
+
 import numpy as np
 import pandas as pd
 from scipy.stats import chi2
 from tabulate import tabulate
 
-from ._utils import dof, stat_poisson_unif
 from ._cv_functions import poisson_cv
-
+from ._utils import dof, stat_poisson_unif
 
 time_decorator = importlib.import_module(
     "QuadratiK.tools._utils"
@@ -252,7 +252,7 @@ class PoissonKernelTest:
 
         if print_fmt == "html":
             summary_string = (
-                "Time taken for execution: {} seconds".format(self.execution_time)
+                f"Time taken for execution: {self.execution_time} seconds"
                 + "<br>Test Results <br>"
                 + f"<br>{self.test_type_} <br>"
                 + tabulate(res, tablefmt=print_fmt, headers=res.columns)
@@ -267,7 +267,7 @@ class PoissonKernelTest:
             )
         else:
             summary_string = (
-                "Time taken for execution: {:.3f} seconds".format(self.execution_time)
+                f"Time taken for execution: {self.execution_time:.3f} seconds"
                 + "\nTest Results \n"
                 + f"{self.test_type_} \n"
                 + tabulate(res, tablefmt=print_fmt, headers=res.columns)
