+++ QuadratiK/kernel_test/__init__.py
@@ -1,4 +1,4 @@
-from ._kernel_test import KernelTest
 from ._h_selection import select_h
+from ._kernel_test import KernelTest
 
 __all__ = ["KernelTest", "select_h"]
