+++ QuadratiK/tools/__init__.py
@@ -1,4 +1,4 @@
-from .graphics import qq_plot, sphere3d, plot_clusters_2d
-from .tools import stats, sample_hypersphere
+from .graphics import plot_clusters_2d, qq_plot, sphere3d
+from .tools import sample_hypersphere, stats
 
 __all__ = ["qq_plot", "sphere3d", "stats", "sample_hypersphere", "plot_clusters_2d"]
