### 2024-09-01 974a4b2
updated black github actions and updated cibuildwheel supported pythons
