+++ pyproject.toml
@@ -49,7 +49,7 @@ pytest-cov = "^4.1.0"
 Homepage = "https://github.com/rmj3197/QuadratiK"
 
 [tool.cibuildwheel]
-skip = "pp*"
+skip = "pp* cp3[0-8]*"
 
 [tool.ruff]
 line-length = 88 # same as black
