+++ .github/workflows/black_check.yml
@@ -1,11 +1,10 @@
-name: Black Formatting Check
+name: Lint
+
 on: [push, pull_request]
+
 jobs:
-  linter_name:
-    name: runner / black formatter
-    runs-on: ubuntu-latest
+  lint:
+    runs-on: macos-latest
     steps:
       - uses: actions/checkout@v4
-      - uses: rickstaa/action-black@v1
-        with:
-          black_args: ". --check"
+      - uses: psf/black@stable
