### 2024-09-01 95186af
removed cibuildwheel
