### 2024-09-01 ba88551
Create CITATION.cff
