### 2024-09-01 b437492
Update CITATION.cff
