+++ README.rst
@@ -111,6 +111,25 @@ License
 
 This project uses the GPL-3.0 license, with a full version of the license included in the repository `here <https://github.com/rmj3197/QuadratiK/blob/master/LICENSE>`_.
 
+
+Citation
+--------
+
+If you use this package, please consider citing it using the following entry:
+
+.. code-block:: tex
+
+    @misc{saraceno2024goodnessoffitclusteringsphericaldata,
+          title={Goodness-of-Fit and Clustering of Spherical Data: the QuadratiK package in R and Python}, 
+          author={Giovanni Saraceno and Marianthi Markatou and Raktim Mukhopadhyay and Mojgan Golzy},
+          year={2024},
+          eprint={2402.02290},
+          archivePrefix={arXiv},
+          primaryClass={stat.CO},
+          url={https://arxiv.org/abs/2402.02290}, 
+    }
+
+
 Funding Information
 ---------------------
 The work has been supported by Kaleida Health Foundation, Food and Drug Administration, and Department of Biostatistics, University at Buffalo. 
