### 2024-09-02 657c9a2
added __slots__
