### 2024-09-02 f68a263
Update _poisson_kernel_test.py
