### 2024-09-02 5a5c3ff
initialized all variables in __init__
