+++ README.rst
@@ -83,7 +83,7 @@ The QuadratiK package is implemented in both **R** and **Python**, providing a c
 
 * **User Interface**: We also provide a dashboard application built using ``streamlit`` allowing users to access the methods implemented in the package without the need for programming.
 
-The **R** implementation is also available on `CRAN <https://cran.r-project.org/web/packages/QuadratiK/index.html>`_.
+The **R** implementation can be found on `CRAN <https://cran.r-project.org/web/packages/QuadratiK/index.html>`_ and the corresponding GitHub repository is available at `QuadratiK in R <https://github.com/giovsaraceno/QuadratiK-package>`_.
 
 Authors
 ---------
