### 2024-09-03 ac42cfa
implemented python typing
