+++ QuadratiK/spherical_clustering/_utils.py
@@ -5,7 +5,7 @@ Contains utility functions for PKBC and PKBD.
 import numpy as np
 
 
-def root_func(x, num_data, alpha_current_h, num_var, mu_denom_h, sum_h_weight_mat):
+def _root_func(x, num_data, alpha_current_h, num_var, mu_denom_h, sum_h_weight_mat):
     """
     Function used for determing the root, used in PKBC.
     """
@@ -16,7 +16,7 @@ def root_func(x, num_data, alpha_current_h, num_var, mu_denom_h, sum_h_weight_ma
     )
 
 
-def c_d_lambda(beta, p, lamda):
+def _c_d_lambda(beta, p, lamda):
     """
     Function used for determining root, used in PKBD.
     """
@@ -28,7 +28,9 @@ def c_d_lambda(beta, p, lamda):
     )
 
 
-def calculate_wcss_euclidean(k, memb_best, dat, mu_best):
+def _calculate_wcss_euclidean(
+    k: int, memb_best: np.ndarray, dat: np.ndarray, mu_best: np.ndarray
+) -> np.ndarray:
     """
     Used for computing the euclidean WCSS for a single cluster.
     """
@@ -36,7 +38,9 @@ def calculate_wcss_euclidean(k, memb_best, dat, mu_best):
     return np.sum(np.linalg.norm(dat[idx] - mu_best[k], axis=1) ** 2)
 
 
-def calculate_wcss_cosine(k, memb_best, dat, mu_best):
+def _calculate_wcss_cosine(
+    k: int, memb_best: np.ndarray, dat: np.ndarray, mu_best: np.ndarray
+) -> np.ndarray:
     """
     Used for computing the cosine WCSS for a single cluster.
     """
