+++ doc/source/api_reference/generated/QuadratiK.poisson_kernel_test.PoissonKernelTest.rst
@@ -73,6 +73,32 @@
     
     
     
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
         PoissonKernelTest.stats
     
     
