+++ doc/source/api_reference/generated/QuadratiK.spherical_clustering.PKBC.rst
@@ -79,6 +79,48 @@
     
     
     
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
+    
         PKBC.fit
     
     
