+++ doc/source/conf.py
@@ -36,6 +36,22 @@ extensions = [
     "myst_parser",
 ]
 
+myst_enable_extensions = [
+    "amsmath",
+    "attrs_inline",
+    "colon_fence",
+    "deflist",
+    "dollarmath",
+    "fieldlist",
+    "html_admonition",
+    "html_image",
+    "replacements",
+    "smartquotes",
+    "strikethrough",
+    "substitution",
+    "tasklist",
+]
+
 templates_path = ["_templates"]
 exclude_patterns = []
 
