+++ doc/source/changelog/v1.2.dev0.rst
@@ -6,7 +6,7 @@ This is the current version under development. So far, this version includes:
     2. **[NEW]** ``__slots__`` are implemented for classes in order to reduce memory footprint.
     3. **[NEW]** All instance variables are now defined in the ``__init__`` method.
     4. **[NEW]** New CI checks for ``Ruff`` and ``Black`` are included.
-    5. **[NEW]** Development Guide has been updated with instructions on using ``POetry``.
+    5. **[NEW]** Development Guide has been updated with instructions on using ``Poetry``.
     6. **[NEW]** The README has been updated with a list of relevant packages in ``R`` and ``Python`` ecosystems.
     7. **[NEW]** Citations have been included in the repository.
     8. **[NEW]** A short guide on Poisson Kernel-based Distributions is included in User Guide.
