+++ doc/source/user_guide/pkbd.rst
@@ -45,6 +45,6 @@ Usage examples are included in `User Guide <user_guide>`_.
 References
 ************
 
-Golzy, M., & Markatou, M. (2020). Poisson Kernel-Based Clustering on the Sphere: Convergence Properties, Identifiability, and a Method of Sampling. Journal of Computational and Graphical Statistics, 29(4), 758–770. https://doi.org/10.1080/10618600.2020.1740713
+Golzy, M., & Markatou, M. (2020). Poisson Kernel-Based Clustering on the Sphere: Convergence Properties, Identifiability, and a Method of Sampling. Journal of Computational and Graphical Statistics, 29(4), 758–770. https://doi.org/10.1080/10618600.2020.1740713.
 
 Sablica, L., Hornik, K., & Leydold, J. (2023). Efficient sampling from the PKBD distribution. Electronic Journal of Statistics, 17(2), 2180-2209.
