+++ QuadratiK/kernel_test/_kernel_test.py
@@ -133,8 +133,8 @@ class KernelTest:
 
     References
     -----------
-        Markatou M., Saraceno G., Chen Y (2023). “Two- and k-Sample Tests Based on Quadratic Distances.
-        ”Manuscript, (Department of Biostatistics, University at Buffalo).
+        Markatou, M., & Saraceno, G. (2024). A unified framework for multivariate two-sample and k-sample 
+        kernel-based quadratic distance goodness-of-fit tests. arXiv preprint arXiv:2407.16374.
 
         Lindsay BG, Markatou M. & Ray S. (2014) Kernels, Degrees of Freedom, and
         Power Properties of Quadratic Distance Goodness-of-Fit Tests, Journal of the American Statistical
