+++ QuadratiK/datasets/data/wine_dataset.rst
@@ -36,7 +36,7 @@ The analysis determined the quantities of 13 constituents found in each of the t
 Source
 ------
 
-Aeberhard,Stefan and Forina,M.. (1991). Wine. UCI Machine Learning Repository. https://doi.org/10.24432/C5PC7J.
+Aeberhard, S. & Forina, M. (1992). Wine [Dataset]. UCI Machine Learning Repository. https://doi.org/10.24432/C5PC7J.
 
 References
 ----------
