+++ QuadratiK/kernel_test/_h_selection.py
@@ -512,7 +512,7 @@ def select_h(
 
     References
     -----------
-        Markatou, M., & Saraceno, G. (2024). A unified framework for multivariate two-sample and 
+        Markatou, M., & Saraceno, G. (2024). A unified framework for multivariate two-sample and
         k-sample kernel-based quadratic distance goodness-of-fit tests. arXiv preprint arXiv:2407.16374.
 
     Examples
