### 2024-09-26 #180 @rmj3197
Dear @acolum, @ab93, and @isabelizimm,

Thank you very much for your insightful review. We apologize for the delay in our response. We have tried to address the points raised in the review below:

# Comments from @acolum

- **This wouldn't be required for final approval, but I think it would be helpful for potential users of your software to list similar packages in the R and Python scientific ecosystems in the README file. For example, what other packages implement similar methods or have similar overall functionality?**

We have added a list of relevant packages in R and Python in the README file.

- **The README file is missing a repo status badge. With the addition of this badge (and potentially others, like the pyOpenSci peer-review badge, in the future), I'd recommend organizing the badges in your README like the badges are organized in this rOpenSci package's README.**

We have added the repo status badge and organized the various other badges according to the categories specified in the example package.

- **Although the README file links to the primary vignette with examples of how to use the package, the other vignettes in the documentation should be linked to here as well.**

We have now linked all vignettes in the README file.

- **Currently, there are a few citations in the bibliography in the README file, but it's not clear which should be cited when citing the software package. More clarity and/or a BibTeX citation is needed. This guidance from GitHub may help.**

We have added a CITATION.cff file in the repository. Additionally, we have also included the BibTex entry in the README.

- **The README file in the root directory of the package's repository has a .rst file extension instead of a .md file extension. Since I'm a first-time reviewer, I'm not sure how important this difference is to pyOpenSci.**

This was clarified by @isabelizimm that a .rst file is fine.

# Comments from @ab93

- **The package uses the Poetry package manager, which is great. As a result, it would be nice to include a small guide to install the package for development using poetry.e.g. poetry install.**

We have now updated the development guide with commands and instructions on using Poetry. Please see the README file for the updated guide.

- **In the package dependencies the Python version requirement is given as a broad python = "^3.9, !=3.9.7".This can cause some downstream applications using this package to break if any of its dependencies do not support a new Python version.So my recommendation will be to only allow Python versions that the package supports, i.e. something likepython = ">=3.9, <3.13"**

This has been updated in the pyproject.toml file with ``[python = ">=3.9, !=3.9.7, <3.13"]``.

- **I see the Black format badge, but I don't see the CI having the black format check. Adding that would be great.**

The black format CI check is included now. The github action can be found at - https://github.com/rmj3197/QuadratiK/actions/workflows/black_check.yml.

- **It would be great to have some linting tools in the CI pipeline. Ruff is a great option, and will eliminate any anti-patterns in the coding practices.**

We have now implemented the Ruff linting. The github action can be found at - https://github.com/rmj3197/QuadratiK/actions/workflows/ruff_linting.yml.

- **Python typing is missing in the code, and would be a very nice to have, e.g. _pkbc.py**

We have now included Python typing. The hints are now being shown in our updated documentation (https://quadratik.readthedocs.io/en/latest/api_reference/index.html).

- **A good practice is also to make sure all instance variables are first defined in the __init__() function, e.g. in the PKBC class,self.dat only gets initialized in fit()**

All instance variables are now first defined in the ``__init__()`` method.

- **Adding ``__slots__`` would be nice, as it reduces memory footprint.**

``__slots__`` are now included. An example is - https://github.com/rmj3197/QuadratiK/blob/master/QuadratiK/kernel_test/_kernel_test.py

- **Raising the base Exception class is not the best practice, a more fine-grained exception raising would be great**

The base Exception class is not raised anymore. We now raise relevant errors.
