### 2024-11-17 d0c3e3e
Minor Fixes
Changed folder doc -> docs, fixed indentation for docstrings.
