### 2024-11-17 aeb15fe
Update _h_selection.py
