### 2024-11-17 d9adfa1
Updated all mentions of doc/ and added Documentation link to pyproject.toml
