### 2024-11-17 f61725d
improved documentation -> added more articles in User Guide
Added additional documentation for h_selection, kernel-based quadratic distances GoF, Poisson kernel-based quadratic distance test of Uniformity on the sphere, and arXiv badge in README
