+++ docs/source/user_guide/index.rst
@@ -16,6 +16,20 @@ Common Usage Examples
 
    basic_usage
 
+Kernel-based quadratic distance (KBQD) Goodness-of-Fit tests
+=============================================================
+.. toctree::
+   :maxdepth: 4
+
+   kbqd.rst
+
+Select the value of the kernel tuning parameter (h)
+====================================================
+.. toctree::
+   :maxdepth: 4
+
+   hselect.rst
+
 Random Sampling from PKBD
 ===========================
 .. toctree::
@@ -23,6 +37,13 @@ Random Sampling from PKBD
 
    pkbd.rst
    gen_plot_rpkb
+
+Poisson kernel-based quadratic distance test of Uniformity on the sphere
+=========================================================================
+.. toctree::
+   :maxdepth: 4
+
+   pktest.rst
 
 Usage Instructions for Dashboard Application
 ==============================================
