### 2024-11-24 98c025a
Bumped version 1.1.2
