### 2024-11-24 763315c
Updated to make code Ruff version 0.8.0 compliant
