+++ QuadratiK/datasets/__init__.py
@@ -4,4 +4,4 @@ from ._dataset import (
     load_wisconsin_breast_cancer_data,
 )
 
-__all__ = ["load_wireless_data", "load_wisconsin_breast_cancer_data", "load_wine_data"]
+__all__ = ["load_wine_data", "load_wireless_data", "load_wisconsin_breast_cancer_data"]
