+++ QuadratiK/kernel_test/_kernel_test.py
@@ -191,38 +191,38 @@ class KernelTest:
     """
 
     __slots__ = (
-        "h",
-        "method",
-        "num_iter",
-        "b",
-        "quantile",
-        "mu_hat",
-        "sigma_hat",
-        "centering_type",
         "alternative",
-        "k_threshold",
-        "random_state",
-        "n_jobs",
-        "x",
-        "y",
-        "test_type_",
-        "execution_time",
-        "un_test_statistic_",
-        "un_cv_",
-        "un_h0_rejected_",
-        "var_un_",
-        "vn_test_statistic_",
-        "vn_cv_",
-        "vn_h0_rejected_",
-        "dn_test_statistic_",
+        "b",
+        "centering_type",
+        "cv_method_",
         "dn_cv_",
         "dn_h0_rejected_",
-        "var_dn_",
-        "trace_test_statistic_",
+        "dn_test_statistic_",
+        "execution_time",
+        "h",
+        "k_threshold",
+        "method",
+        "mu_hat",
+        "n_jobs",
+        "num_iter",
+        "quantile",
+        "random_state",
+        "sigma_hat",
+        "test_type_",
         "trace_cv_",
         "trace_h0_rejected_",
+        "trace_test_statistic_",
+        "un_cv_",
+        "un_h0_rejected_",
+        "un_test_statistic_",
+        "var_dn_",
         "var_trace_",
-        "cv_method_",
+        "var_un_",
+        "vn_cv_",
+        "vn_h0_rejected_",
+        "vn_test_statistic_",
+        "x",
+        "y",
     )
 
     def __init__(
