+++ QuadratiK/tools/__init__.py
@@ -1,4 +1,4 @@
 from .graphics import plot_clusters_2d, qq_plot, sphere3d
 from .tools import sample_hypersphere, stats
 
-__all__ = ["qq_plot", "sphere3d", "stats", "sample_hypersphere", "plot_clusters_2d"]
+__all__ = ["plot_clusters_2d", "qq_plot", "sample_hypersphere", "sphere3d", "stats"]
