### 2024-12-03 #180 @isabelizimm
🎉 YIPPEEE thanks all! `QuadratiK` has been approved by pyOpenSci! Thank you @rmj3197 for submitting and many thanks to @ab93 and @acolum for reviewing this package! 🌹

- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

- [ ] Invite the package maintainer(s) and both reviewers to slack if they wish to join. (let me know if you would like to join!)

---
