### 2024-12-23 5de4358
updated PKBC class to parallelize per init computation
