+++ docs/source/changelog/v1.1.2.rst
@@ -1,8 +1,6 @@
 QuadratiK Version 1.1.2
 ========================
 
-This is the current version under development. So far, this version includes:
-
 1. **[NEW]** Code fixed to raise fine-grained exceptions instead of the base `Exception`.
 2. **[NEW]** ``__slots__`` are implemented for classes to reduce memory footprint.
 3. **[NEW]** All instance variables are now defined in the ``__init__`` method.
