### 2024-12-23 0b9919c
updated README to include pyOpenSci badge
