### 2024-12-23 5c7b8bd
Update README.rst
Added Zenodo Badge
