### 2024-12-23 #180 @rmj3197
Hello @isabelizimm,

Thank you very much for approving the package!

I wanted to let you know that I have released a new version and completed the Zenodo tagging (and watching) as well as the survey.

We are currently working on the blog post and will submit a separate pull request for it, referencing this issue once it's ready.

I would also be interested in joining the Slack channel.

Thanks again for your help and support, and wishing you a happy holiday season!
