### 2025-02-07 2f43315
uncommented code for h-selection example, and experimental try to enable plotly on RTD
