+++ docs/source/conf.py
@@ -63,7 +63,9 @@ html_theme = "sphinx_book_theme"
 html_static_path = ["_static"]
 html_css_files = ["css/custom.css"]
 # needed for plotly, see https://myst-nb.readthedocs.io/en/latest/render/interactive.html#plotly
-html_js_files = ["https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.4/require.min.js"]
+html_js_files = [
+    "https://cdnjs.cloudflare.com/ajax/libs/require.js/2.3.4/require.min.js"
+]
 strip_signature_backslash = True
 
 # -- Options for Latex output -------------------------------------------------
