### 2025-02-07 3ff58ba
Revert "Update requirements.txt"
This reverts commit 28dff6b4343b6385d84c146ab5cd68db51ce61d6.
