### 2025-02-07 9543dc9
Revert "uncommented code for h-selection example, and experimental try to enable plotly on RTD"
This reverts commit 2f4331513403e31334cfae0be859c45cb4595106.
