### 2025-02-07 baee3f9
changes to Dashboard and conf.py
