+++ QuadratiK/ui/pages/2_Two_Sample_Test.py
@@ -21,24 +21,76 @@ st.title("Two Sample Test")
 
 st.write("Performs the Nonparametric Two Sample Test")
 
-with st.expander("Click to view code"):
+with st.expander("Click to view example code in Python and R"):
     code_python = """
+    import numpy as np
+
+    np.random.seed(0)
+    from scipy.stats import skewnorm
+
     from QuadratiK.kernel_test import KernelTest
-    X = Read your data file here
-    Y = Read your data file here
-    two_sample_test = KernelTest(h = 0.5).test(X,Y)
-    two_sample_test.summary()
+
+    # data generation
+    X_2 = np.random.multivariate_normal(mean=np.zeros(4), cov=np.eye(4), size=200)
+    Y_2 = skewnorm.rvs(
+    size=(200, 4),
+    loc=np.zeros(4),
+    scale=np.ones(4),
+    a=np.repeat(0.5, 4),
+    random_state=20,
+    )
+    # performing the two sample test
+    two_sample_test = KernelTest(h=2, num_iter=150, random_state=42).test(X_2, Y_2)
+
+    # printing the summary for the two sample test
+    print(two_sample_test.summary())
     """
     st.code(code_python, language="python")
 
     code_R = """
-    library(QuadratiK)
-    X = Read your data file here
-    Y = Read your data file here
-    two_test <- kb.test(x=X, y=Y, h=2)
-    summary(two_test)
+    library(sn)         # For generating skew-normal distributed data
+    library(mvtnorm)    # For generating multivariate normal data
+    library(QuadratiK)  
+
+    # Set parameters
+    n <- 100           # Number of samples
+    d <- 4             # Dimension of the data
+    skewness_y <- 0.5  # Skewness parameter for Y distribution
+
+    # Set seed for reproducibility
+    set.seed(2468)
+
+    # Generate multivariate normal data for X
+    x_2 <- rmvnorm(n, mean = rep(0, d))  # Mean vector of zeros, identity covariance
+
+    # Generate skew-normal data for Y
+    y_2 <- rmsn(n = n, xi = 0, Omega = diag(d), alpha = rep(skewness_y, d))  
+
+    # Perform a statistical test to compare the two datasets
+    two_test <- kb.test(x = x_2, y = y_2, h = 2)
+
+    # Output the test result
+    two_test
     """
     st.code(code_R, language="r")
+
+st.subheader("Input Instructions", divider="grey")
+st.write("1. Upload the data file in .txt or .csv format for both the X and Y datasets.")
+st.write(
+    "2. The file may contain a header (see image below for reference). If headers are present, check the box. The checkbox is selected by default. Please ensure that both X and Y either contain headers or neither contain headers."
+)
+st.write("3. Specify the separator or delimiter used in both the X and Y datasets; the default is a comma (,).")
+st.write("4. Once the data files are uploaded, specify the values of bandwidth parameter, proportion of subsampling samples to be used, and number of iterations for critical value estimation. Default values are provided.")
+
+st.image(
+    str(
+        importlib.resources.files("QuadratiK.ui").joinpath(
+            "pages/assets/two_sample_test_format.png"
+        )
+    ),
+    caption="Sample data format for two sample test",
+    use_container_width=True,
+)
 
 delim = st.text_input("**Enter the delimiter**", ",")
 header_exist = st.checkbox(
