+++ QuadratiK/__init__.py
@@ -1,6 +1,6 @@
 from importlib import import_module
 
-__version__ = "1.1.2"
+__version__ = "1.1.3dev0"
 
 submodules = [
     "kernel_test",
