### 2025-02-07 535edaf
trying to load plotly on RTD
