### 2025-02-07 4302446
removed output for ipynb files
