+++ .readthedocs.yml
@@ -8,7 +8,7 @@ version: 2
 build:
   os: ubuntu-20.04
   tools:
-    python: '3.9'
+    python: '3.12'
 
 # Build documentation in the docs/ directory with Sphinx
 sphinx:
