+++ docs/source/conf.py
@@ -36,6 +36,8 @@ extensions = [
     "myst_parser",
 ]
 
+nbsphinx_execute = "always"
+
 myst_enable_extensions = [
     "amsmath",
     "attrs_inline",
