### 2025-02-08 4a66244
added more instructions to dashboard
