+++ docs/requirements.txt
@@ -4,4 +4,5 @@ sphinx-autobuild
 sphinx-book-theme
 nbsphinx
 ipykernel
+kaleido
 git+https://github.com/rmj3197/QuadratiK.git@dev-acm
