### 2025-02-18 7cfdf76
Changes to dashboard and plotting behavior
Removed nest-asyncio, changed the dashboard deployment using async to runpy. Fixed plot behavior - interactive vs non-interactive environments.
