+++ QuadratiK/tools/graphics.py
@@ -1,3 +1,4 @@
+import sys
 from typing import Optional, Union
 
 import matplotlib.pyplot as plt
@@ -224,5 +225,8 @@ def plot_clusters_2d(
     plt.plot(unit_circle_x, unit_circle_y, linestyle="dashed", color="red")
     plt.xlabel("Feature 1")
     plt.ylabel("Feature 2")
-    plt.close()
+    if "IPython" in sys.modules:
+        plt.close()
+    else:
+        plt.show()
     return fig
