+++ QuadratiK/tools/_utils.py
@@ -1,3 +1,4 @@
+import sys
 import time
 from functools import wraps
 from typing import Any, Callable, TypeVar, Union
@@ -55,7 +56,11 @@ def _qq_plot_twosample(
         axes[col].set_ylabel("Q2")
         fig.suptitle("QQ Plots", fontsize=16)
     fig.subplots_adjust(hspace=0.5)
-    plt.close()
+
+    if "IPython" in sys.modules:
+        plt.close()
+    else:
+        plt.show()
     return fig
 
 
@@ -111,7 +116,10 @@ def _qq_plot_onesample(
         axes[col].set_ylabel("Sample Quantiles")
         fig.suptitle("QQ Plots", fontsize=16)
     fig.subplots_adjust(hspace=0.5)
-    plt.close()
+    if "IPython" in sys.modules:
+        plt.close()
+    else:
+        plt.show()
     return fig
 
 
