+++ pyproject.toml
@@ -21,7 +21,6 @@ scikit-learn = "^1.3"
 scipy = "^1.11"
 streamlit = "^1.30.0"
 matplotlib = "^3.8.2"
-nest-asyncio = "^1.5"
 plotly = "^5.15.0"
 tabulate = "^0.8"
 pandas = "^2.1.3"
