### 2025-02-18 b19e413
changes to basic_usage.py and conf.py
