+++ docs/source/conf.py
@@ -62,6 +62,9 @@ exclude_patterns = []
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
 html_theme = "sphinx_book_theme"
+html_theme_options = {
+    "use_download_button": True,
+}
 html_static_path = ["_static"]
 html_css_files = ["css/custom.css"]
 strip_signature_backslash = True
