### 2025-02-20 e9a7874
modified code to return plots and modified basic_usage.ipynb
