+++ .readthedocs.yml
@@ -19,7 +19,8 @@ sphinx:
 #  configuration: mkdocs.yml
 
 # Optionally build your docs in additional formats such as PDF and ePub
-formats: all
+formats:
+  - pdf
 
 # Optionally set the version of Python and requirements required to build your docs
 python:
