+++ .github/workflows/codecov.yml
@@ -3,23 +3,27 @@ name: Codecov Workflow
 on: [push, pull_request]
 
 jobs:
-  build:
-    runs-on: macos-latest
-    name: Run pytest with cov
+  test:
+    strategy:
+      matrix:
+        python-version: ['3.9', '3.10', '3.11', '3.12', '3.13']
+        os: [windows-latest, ubuntu-latest, macos-latest]
+    runs-on: ${{ matrix.os }}
+    name: Python ${{ matrix.python-version }} on ${{ matrix.os }}
     steps:
     - uses: actions/checkout@v4
-    - name: Setup Python 3.12
-      uses: actions/setup-python@v2
+    - name: Setup Python ${{ matrix.python-version }}
+      uses: actions/setup-python@v5.4.0
       with:
-        python-version: 3.12.0
+        python-version: ${{ matrix.python-version }}
     - name: Install poetry
       run: pip install poetry
     - name: Install all dependencies along with test dependencies with poetry
       run: poetry install --with tests
     - name: Run tests and collect coverage
-      run: poetry run pytest --cov-report=xml --cov . 
+      run: poetry run pytest --cov-report=xml --cov .
     - name: Upload coverage reports to Codecov
-      uses: codecov/codecov-action@v3.1.6
+      uses: codecov/codecov-action@v5.3.1
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         slug: rmj3197/QuadratiK
