+++ pyproject.toml
@@ -24,7 +24,10 @@ matplotlib = "^3.8.2"
 plotly = "^5.15.0"
 tabulate = "^0.8"
 pandas = "^2.1.3"
-numpy = "^1.26.2"
+numpy = [
+    {version = "^1.26.2", python = ">=3.9,<3.13"},
+    {version = "^2.0.0", python = ">=3.13"}
+]
 
 [build-system]
 requires = ["poetry-core", "setuptools"]
