### 2025-02-23 fc9a8b4
updated dashboard documentation to reflect new changes and made changelog visible.
