+++ QuadratiK/ui/pages/3_K_Sample_Test.py
@@ -135,7 +135,7 @@ if data is not None:
         "Enter a value of number of iterations to be used for critical value estimation",
         value=500,
     )
-    h = st.number_input("Enter a value of tuning parameter h", value=1)
+    h = st.number_input("Enter a value of tuning parameter h", value=1.5)
     b = st.number_input(
         "Enter a value for the proportion of subsampling samples to be used", value=0.9
     )
