+++ .github/workflows/codecov.yml
@@ -6,14 +6,14 @@ jobs:
   test:
     strategy:
       matrix:
-        python-version: ['3.9.21', '3.10.16', '3.11.11', '3.12.9', '3.13.2']
+        python-version: ['3.9', '3.10', '3.11', '3.12', '3.13']
         os: [windows-latest, ubuntu-latest, macos-latest]
     runs-on: ${{ matrix.os }}
     name: Python ${{ matrix.python-version }} on ${{ matrix.os }}
     steps:
     - uses: actions/checkout@v4
     - name: Setup Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v3
+      uses: actions/setup-python@v2
       with:
         python-version: ${{ matrix.python-version }}
     - name: Install poetry
@@ -23,7 +23,7 @@ jobs:
     - name: Run tests and collect coverage
       run: poetry run pytest --cov-report=xml --cov .
     - name: Upload coverage reports to Codecov
-      uses: codecov/codecov-action@v5.3.1
+      uses: codecov/codecov-action@v3.1.6
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         slug: rmj3197/QuadratiK
