+++ .github/workflows/codecov.yml
@@ -7,15 +7,17 @@ jobs:
     strategy:
       matrix:
         python-version: ['3.9', '3.10', '3.11', '3.12', '3.13']
-        os: [windows-2022, ubuntu-latest, macos-14]
+        os: [windows-latest, ubuntu-latest, macos-latest]
     runs-on: ${{ matrix.os }}
     name: Python ${{ matrix.python-version }} on ${{ matrix.os }}
     steps:
     - uses: actions/checkout@v4
     - name: Setup Python ${{ matrix.python-version }}
-      uses: actions/setup-python@v2
+      uses: actions/setup-python@v5.4.0
       with:
         python-version: ${{ matrix.python-version }}
+    - name: Print Python version
+      run: python --version
     - name: Install poetry
       run: pip install poetry
     - name: Install all dependencies along with test dependencies with poetry
@@ -23,7 +25,7 @@ jobs:
     - name: Run tests and collect coverage
       run: poetry run pytest --cov-report=xml --cov .
     - name: Upload coverage reports to Codecov
-      uses: codecov/codecov-action@v3.1.6
+      uses: codecov/codecov-action@v5.3.1
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         slug: rmj3197/QuadratiK
