+++ .github/workflows/codecov.yml
@@ -23,7 +23,8 @@ jobs:
     - name: Install all dependencies along with test dependencies with poetry
       run: poetry install --with tests
     - name: Run tests and collect coverage
-      run: poetry run pytest --cov-report=xml --cov .
+      # run: poetry run pytest --cov-report=xml --cov .
+      run: pytest --cov-report xml:cov.xml --cov=. tests
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v5.3.1
       with:
