### 2025-02-27 #180 @lwasser
HI everyone!! This issue is still open, but I will close it!

✅  I've updated the header info with the Zenodo doi, version accepted, and date accepted. @rmj3197 I see that you are already in our Slack. If anyone else (Reviewers maintainers etc) involved in this issue would like to join our slack, please email or email me at leah at pyopensci.org!!

thank you again for submitting a blog post about quadratik @rmj3197
